requirements_test.txt
lemmy==2.1.0
spacy==2.3.2
pandas==1.1.3
networkx==2.5
matplotlib==3.3.2
pytest==6.1.1
watchdog==0.10.3
python-dotenv==0.14.0
rdflib==5.0.0